pyproject.toml
[tool.poetry]
name = "desk-python-sdk"
version = "0.0.0"
description = "Desk Python SDK"
authors = ["chenciao8 <chenciao8@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/chenciao8/desk-python-sdk"
repository = "https://github.com/chenciao8/desk-python-sdk"
documentation = "https://desk-python-sdk.readthedocs.io"
packages = [
    { include = "desk", from = "src" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
]

[tool.poetry.urls]
Changelog = "https://github.com/chenciao8/desk-python-sdk/releases"

[tool.poetry.dependencies]
python = "^3.9"
click = ">=8.0.1"
docutils = "^0.21.2"
eth-account = "^0.13.4"
web3 = "^7.7.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
Pygments = ">=2.10.0"
black = ">=21.10b0"
coverage = {extras = ["toml"], version = ">=6.2"}
darglint = ">=1.8.1"
flake8 = ">=4.0.1"
flake8-bandit = ">=2.1.2"
flake8-bugbear = ">=21.9.2"
flake8-docstrings = ">=1.6.0"
flake8-rst-docstrings = ">=0.2.5"
furo = ">=2021.11.12"
isort = ">=5.10.1"
mypy = ">=0.930"
pep8-naming = ">=0.12.1"
pre-commit = ">=2.16.0"
pre-commit-hooks = ">=4.1.0"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"
safety = ">=1.10.3"
sphinx = ">=4.3.2"
sphinx-autobuild = ">=2021.3.14"
sphinx-click = ">=3.0.2"
typeguard = ">=2.13.3"
xdoctest = {extras = ["colors"], version = ">=0.15.10"}
myst-parser = {version = ">=0.16.1"}

[tool.poetry.scripts]
desk-python-sdk = "desk.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["desk", "tests"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "src"
